Add AbstractCommonProtocolMethodProcessor to share some common methods. (#325)
(cherry picked from commit 6e0f76774338923563445ce6a4b118d07e7dd314)
